refactor: Move kanji drawer away from kanjilist component
